requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.5.0
billiard==3.5.0.5
celery==4.0.2
certifi==2021.10.8
chardet==3.0.4
Django==2.1.3
django-celery==3.2.2
django-crontab==0.7.1
djangorestframework==3.12.2
flower==0.9.7
greenlet==1.1.2
humanize==3.14.0
idna==2.6
importlib-metadata==4.10.1
Jinja2==3.0.3
kombu==4.6.11
MarkupSafe==2.0.1
mysqlclient==2.0.1
numpy==1.21.5
pandas==1.1.5
Pillow==8.0.1
prettytable==3.0.0
prometheus-client==0.8.0
pyecharts==1.9.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
redis==2.10.6
requests==2.18.4
semantic-version==2.8.5
setuptools-rust==1.1.2
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
tornado==6.1
typing-extensions==4.0.1
urllib3==1.22
vine==1.3.0
wcwidth==0.2.5
zipp==3.7.0
